Move DownloadContainer to Download. Create downloadFileInNewTab util fn. Open download sliding panel on left.
